# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS ESTIMATION BACKEND DOCUMENTATION",
    "description": "Our story started in 2015 when there was a need of Construction database for helping bidders to bid accurately with reliable information's and estimates in the Rwanda construction industry. Continuous determination, hard work and passion have been constant themes throughout our history and present. As Rwanda Construction industry becomes increasingly important on the global stage, it's our believe that, with our own innovative mindset, we can make the construction industry a better one for national development. So we challenge ourselves to evolve, innovate and experiment to make work easier and faster for everyone as the leading local software brand in the Rwanda construction industry. We are currently the quantities and estimations software dealers which are made in Africa its self initially from Ghana and same technology has been customized to fit Rwanda construction industry and we are super proud to start the race with national development in minds. Today we are privileged to connect with thousands of Engineers and other construction related professionals who do daily estimates or quantities in the construction industry.",
    "version": "1.0.0",
    "contact": {
      "name": "DS ESTIMATION",
      "url": "http://dsestimation.com",
      "email": "dsestimation@gmail.com"
    }
  },
  "host": "dse-estimation-backend.herokuapp.com",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ],
  "paths": {
...
    "/owner/equipment/create": {
      "post": {
        "tags": ["Default Equipment Routes"],
        "summary": "Create an equipment",
        "description": "Create an equipment",
        "operationId": "createEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Equipment object that needs to be created equipment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment created successfully"
          },
          "400": {
            "description": "Some equipment field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/equipment/{id}/update": {
      "patch": {
        "tags": ["Default Equipment Routes"],
        "summary": "Update an equipment",
        "description": "Update an equipment",
        "operationId": "updateEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the equipment to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Equipment object that needs to be updated default equipment, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment updated successfully"
          },
          "400": {
            "description": "Some equipment field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Equipment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/equipment/{id}/delete": {
      "delete": {
        "tags": ["Default Equipment Routes"],
        "summary": "Delete an equipment",
        "description": "Delete an equipment",
        "operationId": "deleteEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the equipment to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Equipment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/equipment/create": {
      "post": {
        "tags": ["Client Equipment Routes"],
        "summary": "Create an equipment",
        "description": "Create an equipment",
        "operationId": "createEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Equipment object that needs to be created equipment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment created successfully"
          },
          "400": {
            "description": "Some equipment field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/equipment/{id}/update": {
      "patch": {
        "tags": ["Client Equipment Routes"],
        "summary": "Update an equipment",
        "description": "Update an equipment",
        "operationId": "updateEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the equipment to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Equipment object that needs to be updated Client equipment, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Equipment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment updated successfully"
          },
          "400": {
            "description": "Some equipment field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Equipment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/equipment/{id}/delete": {
      "delete": {
        "tags": ["Client Equipment Routes"],
        "summary": "Delete an equipment",
        "description": "Delete an equipment",
        "operationId": "deleteEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the equipment to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Equipment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/supplier/all": {
      "get": {
        "tags": ["Client Supplier Routes"],
        "summary": "Get all suppliers",
        "description": "Get all suppliers",
        "operationId": "getAllSuppliers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Suppliers Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/supplier/create": {
      "post": {
        "tags": ["Client Supplier Routes"],
        "summary": "Create a supplier",
        "description": "Create a supplier",
        "operationId": "createSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Supplier object that needs to be created supplier",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier created successfully"
          },
          "400": {
            "description": "Some supplier field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/supplier/{id}/update": {
      "patch": {
        "tags": ["Client Supplier Routes"],
        "summary": "Update a supplier",
        "description": "Update a supplier",
        "operationId": "updateSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the supplier to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Supplier object that needs to be updated supplier",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier updated successfully"
          },
          "400": {
            "description": "Some supplier field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Supplier not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/supplier/{id}/delete": {
      "delete": {
        "tags": ["Client Supplier Routes"],
        "summary": "Delete a supplier",
        "description": "Delete a supplier",
        "operationId": "deleteSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the supplier to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Supplier not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/owner/equipment/{id}/approved": {
      "patch": {
        "tags": ["Unverified data Routes"],
        "summary": "Approve equipment",
        "description": "Approve equipment",
        "operationId": "approveEquipment",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the equipment to approve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment approved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Equipment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/global/supplier/all": {
      "get": {
        "tags": ["Global Routes"],
        "summary": "Get all suppliers from our database",
        "description": "Get all suppliers from our database",
        "operationId": "getAllSuppliers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Suppliers Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Equipment": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "caveragePerUnit": {
          "type": "number"
        },
        "unit": {
          "type": "string"
        },
        "hireRatePrice": {
          "type": "number"
        },
        "number": {
          "type": "number"
        },
        "supplierId": {
          "type": "number"
        },
        "brand": {
          "type": "string"
        },
        "inputUnit": {
          "type": "string"
        },
        "inputQuantity": {
          "type": "number"
        },
        "outputUnit": {
          "type": "string"
        },
        "outputQuantity": {
          "type": "number"
        },
        "adjustFactor": {
          "type": "number"
        }
      }
    },
...
    "Supplier": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "projectId": {
          "type": "number"
        }
      }
    },
...
  }
}